[vchanmanager] Always process timesync request in message proxy
Initial idea was to process it only during provisioning as we don't have CM
at this time. But, after provisioning CM expects node configuration as first
message and returns error if it receives clock sync request. It should be
solved on CM side. But as temporary solution it will be handled in message
proxy.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>